# perf: Add optimization to re-write batched insert stataments.