--- a/pgjdbc/src/main/java/org/postgresql/core/NativeQuery.java
+++ b/pgjdbc/src/main/java/org/postgresql/core/NativeQuery.java
@@ -19,6 +19,7 @@ public class NativeQuery {
 
   public final String nativeSql;
   public final int[] bindPositions;
+  public final boolean isBatchedReWriteCompatible;
 
   static {
     for (int i = 1; i < BIND_NAMES.length; i++) {
@@ -27,13 +28,14 @@ public class NativeQuery {
   }
 
   public NativeQuery(String nativeSql) {
-    this(nativeSql, NO_BINDS);
+    this(nativeSql, NO_BINDS, false);
   }
 
-  public NativeQuery(String nativeSql, int[] bindPositions) {
+  public NativeQuery(String nativeSql, int[] bindPositions, boolean compatible) {
     this.nativeSql = nativeSql;
     this.bindPositions =
         bindPositions == null || bindPositions.length == 0 ? NO_BINDS : bindPositions;
+    this.isBatchedReWriteCompatible = compatible;
   }
 
   /**
@@ -75,5 +77,9 @@ public class NativeQuery {
    */
   public static String bindName(int index) {
     return index < BIND_NAMES.length ? BIND_NAMES[index] : "$" + index;
+  }
+
+  public static int bindsCount() {
+    return BIND_NAMES.length;
   }
 }