--- a/pgjdbc/src/main/java/org/postgresql/core/ParameterList.java
+++ b/pgjdbc/src/main/java/org/postgresql/core/ParameterList.java
@@ -172,17 +172,11 @@ public interface ParameterList {
   String toString(int index);
 
   /**
-   * Overwrite current parameters with parameters in the provided list.
-   * Current parameters are cleared.
-   * @param list of parameters to overwrite with.
-   */
-  void addAll(ParameterList list);
-
-  /**
    * Use this operation to append more parameters to the current list.
    * @param list of parameters to append with.
+   * @throws SQLException fault raised if driver or back end throw an exception
    */
-  void appendAll(ParameterList list);
+  void appendAll(ParameterList list) throws SQLException ;
 
   /**
    * Returns the bound parameter values.