--- a/pgjdbc/src/main/java/org/postgresql/core/Parser.java
+++ b/pgjdbc/src/main/java/org/postgresql/core/Parser.java
@@ -50,6 +50,8 @@ public class Parser {
     StringBuilder nativeSql = new StringBuilder(query.length() + 10);
     List<Integer> bindPositions = null; // initialized on demand
     List<NativeQuery> nativeQueries = null;
+    boolean isCurrentReWriteCompatible = false;
+    int afterValuesParens = 0;
 
     boolean whitespaceOnly = true;
     for (int i = 0; i < aChars.length; ++i) {
@@ -79,6 +81,10 @@ public class Parser {
 
         case '(':
           inParen++;
+          afterValuesParens++;
+          if (afterValuesParens > 1) {
+            isCurrentReWriteCompatible = false;
+          }
           break;
 
         case ')':
@@ -117,13 +123,44 @@ public class Parser {
                 nativeQueries = new ArrayList<NativeQuery>();
               }
 
-              nativeQueries.add(new NativeQuery(nativeSql.toString(), toIntArray(bindPositions)));
+              nativeQueries.add(new NativeQuery(nativeSql.toString(), toIntArray(bindPositions),isCurrentReWriteCompatible));
             }
             // Prepare for next query
             if (bindPositions != null) {
               bindPositions.clear();
             }
             nativeSql.setLength(0);
+            isCurrentReWriteCompatible = false;
+          }
+          break;
+
+        case 'i':
+          isCurrentReWriteCompatible = isCurrentReWriteCompatible || (Parser.parseInsertKeyword(aChars, i, false));
+          break;
+
+        case 'I':
+          isCurrentReWriteCompatible = isCurrentReWriteCompatible || (Parser.parseInsertKeyword(aChars, i, true));
+          break;
+
+        case 'r':
+          // exclude insert statements with returning keyword
+          isCurrentReWriteCompatible = isCurrentReWriteCompatible && (Parser.parseReturningKeyword(aChars, i, false) == false);
+          break;
+
+        case 'R':
+          // exclude insert statements with RETURNING keyword
+          isCurrentReWriteCompatible = isCurrentReWriteCompatible && (Parser.parseReturningKeyword(aChars, i, true) == false);
+          break;
+
+        case 'v':
+          if (Parser.parseValuesKeyword(aChars, i, false)) {
+            afterValuesParens = 0 ;
+          }
+          break;
+
+        case 'V':
+          if (Parser.parseValuesKeyword(aChars, i, true)) {
+            afterValuesParens = 0 ;
           }
           break;
 
@@ -140,7 +177,7 @@ public class Parser {
       return nativeQueries != null ? nativeQueries : Collections.<NativeQuery>emptyList();
     }
 
-    NativeQuery lastQuery = new NativeQuery(nativeSql.toString(), toIntArray(bindPositions));
+    NativeQuery lastQuery = new NativeQuery(nativeSql.toString(), toIntArray(bindPositions), isCurrentReWriteCompatible);
 
     if (nativeQueries == null) {
       return Collections.singletonList(lastQuery);
@@ -336,6 +373,66 @@ public class Parser {
       }
     }
     return offset;
+  }
+
+  /**
+   * Parse string to check presence of INSERT keyword.
+   * @param query char[] of the query statement
+   * @param offset position of query to start checking
+   * @param isUpper is the text expected to be upper/lower case
+   * @return boolean indicates presence of word
+   */
+  public static boolean parseInsertKeyword(final char[] query, int offset, boolean isUpper) {
+    if (query.length < (offset + 7)) {
+      return false;
+    }
+
+    if (isUpper && query[offset] == 'I' && query[offset + 1] == 'N' && query[offset + 2] == 'S' && query[offset + 3] == 'E' && query[offset + 4] == 'R' && query[offset + 5] == 'T' ) {
+      return true;
+    } else if ( !isUpper &&  query[offset] == 'i' &&  query[offset + 1] == 'n' && query[offset + 2] == 's' && query[offset + 3] == 'e' && query[offset + 4] == 'r' && query[offset + 5] == 't') {
+      return true;
+    }
+    return false;
+  }
+
+  /**
+   * Parse string to check presence of RETURNING keyword
+   * @param query char[] of the query statement
+   * @param offset position of query to start checking
+   * @param isUpper is the text expected to be upper/lower case
+   * @return boolean indicates presence of word
+   */
+  public static boolean parseReturningKeyword(final char[] query, int offset, boolean isUpper) {
+    if (query.length < (offset + 9)) {
+      return false;
+    }
+
+    if ( isUpper && query[offset] == 'R' && query[offset + 1] == 'E' && query[offset + 2] == 'T' && query[offset + 3] == 'U' && query[offset + 4] == 'R' && query[offset + 5] == 'N' && query[offset + 6] == 'I' && query[offset + 7] == 'N' && query[offset + 8] == 'G' ) {
+      return true;
+    } else if ( !isUpper && query[offset] == 'r' && query[offset + 1] == 'e' && query[offset + 2] == 't' && query[offset + 3] == 'u' && query[offset + 4] == 'r' && query[offset + 5] == 'n' && query[offset + 6] == 'i' && query[offset + 7] == 'n' && query[offset + 8] == 'g' ) {
+      return true;
+    }
+    return false;
+  }
+
+  /**
+   * Parse string to check presence of VALUES keyword
+   * @param query char[] of the query statement
+   * @param offset position of query to start checking
+   * @param isUpper is the text expected to be upper/lower case
+   * @return boolean indicates presence of word
+   */
+  public static boolean parseValuesKeyword(final char[] query, int offset, boolean isUpper) {
+    if (query.length < (offset + 6)) {
+      return false;
+    }
+
+    if ( isUpper && query[offset] == 'V' && query[offset + 1] == 'A' && query[offset + 2] == 'L' && query[offset + 3] == 'U' && query[offset + 4] == 'E' && query[offset + 5] == 'S' ) {
+      return true;
+    } else if ( !isUpper && query[offset] == 'v' && query[offset + 1] == 'a' && query[offset + 2] == 'l' && query[offset + 3] == 'u' && query[offset + 4] == 'e' && query[offset + 5] == 's' ) {
+      return true;
+    }
+    return false;
   }
 
   /**