… pgjdbc/src/main/java/org/postgresql/core/QueryExecutor.java
… pgjdbc/src/main/java/org/postgresql/core/v2/FastpathParameterList.java
… pgjdbc/src/main/java/org/postgresql/core/v2/SimpleParameterList.java
… pgjdbc/src/main/java/org/postgresql/core/v2/V2Query.java
… pgjdbc/src/main/java/org/postgresql/core/v3/BatchedQueryDecorator.java
… pgjdbc/src/main/java/org/postgresql/core/v3/CompositeParameterList.java
… pgjdbc/src/main/java/org/postgresql/core/v3/CompositeQuery.java
… pgjdbc/src/main/java/org/postgresql/core/v3/QueryExecutorImpl.java
… pgjdbc/src/main/java/org/postgresql/core/v3/SimpleParameterList.java
… pgjdbc/src/main/java/org/postgresql/core/v3/SimpleQuery.java
… pgjdbc/src/main/java/org/postgresql/core/v3/V3ParameterList.java
… pgjdbc/src/main/java/org/postgresql/ds/common/BaseDataSource.java
… pgjdbc/src/main/java/org/postgresql/jdbc/CachedQueryCreateAction.java
… pgjdbc/src/main/java/org/postgresql/jdbc/PgConnection.java
… pgjdbc/src/main/java/org/postgresql/jdbc/PgPreparedStatement.java
… pgjdbc/src/main/java/org/postgresql/jdbc/PgStatement.java
… pgjdbc/src/test/java/org/postgresql/core/v2/V2ParameterListTests.java
… pgjdbc/src/test/java/org/postgresql/core/v3/V3ParameterListTests.java
… pgjdbc/src/test/java/org/postgresql/jdbc/DeepBatchedInsertStatementTest.java
… pgjdbc/src/test/java/org/postgresql/test/core/v2/V2ParameterListTests.java
… pgjdbc/src/test/java/org/postgresql/test/core/v3/V3ParameterListTests.java
… pgjdbc/src/test/java/org/postgresql/test/jdbc2/BatchExecuteTest.java
… pgjdbc/src/test/java/org/postgresql/test/jdbc2/BatchedInsertDoubleRowInSingleBatch.java
… pgjdbc/src/test/java/org/postgresql/test/jdbc2/BatchedInsertReWriteEnabledTest.java
… pgjdbc/src/test/java/org/postgresql/test/jdbc2/BatchedInsertStatementPreparingTest.java
… pgjdbc/src/test/java/org/postgresql/test/jdbc2/DeepBatchedInsertStatementTest.java
… pgjdbc/src/test/java/org/postgresql/test/jdbc2/Jdbc2TestSuite.java
… pgjdbc/src/test/java/org/postgresql/test/jdbc3/CompositeQueryParseTest.java