Hi,
 This PR will add the ability of the driver to re-write insert statements. When the insert statement meets basic conditions.
 This will improve throughput performance.

Jeremy