@vlsi 2016-01-14
>This PR will add the ability of the driver to re-write insert statements

Can you elaborate more which inserts are rewritten and what is the rewrite result?

>This will improve throughput performance.

Could you please add performance tests to justify the point?