@whitingjr 2016-01-14
@vlsi Can you elaborate more which inserts are rewritten and what is the rewrite result?
 Simple INSERT statements. Take this statement
INSERT INTO testbatch VALUES (?,?)

called x2 times using addBatch()
will be collapsed into this form
INSERT INTO testbatch VALUES (?,?),(?,?)

Excluded from the optimization are INSERT statements with RETURNING keyword. Or when it already has multi VALUES row blocks. 

@vlsi Could you please add performance tests to justify the point?
This thread has performance results comparing batching of individual versus re-written statement.
http://www.postgresql.org/message-id/55130DC8.2070508@redhat.com