@vlsi 2016-01-14
Can you execute the same benchmark two more times?
1) Against current 9.4.1208-SNAPSHOT
2) Against 1208+your patch

Bonus question: have you investigated **why** multivalues improve performance?
Is it something that can be reproduced with pgbench, for example?
In other words, if there is a bottleneck in pgjdbc batch processing, then it might be better to resolve it in the first place.

PS. I like the following options more than "ops/sec".
```
@BenchmarkMode(Mode.AverageTime)
@OutputTimeUnit(TimeUnit.MILLISECONDS)
```