@whitingjr 2016-01-14
> Can you execute the same benchmark two more times?
The benchmark uses pure SQL. pgjdbc is not used to push the workload.
> In other words, if there is a bottleneck in pgjdbc batch processing, then it might be better to resolve it in the first place.
 Given pgjdbc is not involved batch processing doesn't need fixing.