@whitingjr 2016-01-14
> have you investigated why multivalues improve performance?

 I've found some discussions on stackexchange.com on this topic. Take a look at this [thread](https://dba.stackexchange.com/questions/76334/multi-row-insert-vs-multiple-single-row-inserts)