@vlsi 2016-01-14
>The benchmark uses pure SQL. pgjdbc is not used to push the workload.

You claim to improve performance.
I want to see measurement results that would highlight that.
Does that make sense?

Your charts in the link above show *potential* improvement.
However, there is no measurements on:
1) Performance improvement that the fix provides
2) Overhead that it creates. I believe overhead should be minimal, but performance issues are hard to reason about.