@vlsi 2016-01-14
@whitingjr , here's what I mean.
It looks like your code degrades as the number of rows increases.
Can you please look into it?

I use existing benchmark for batch inserts: https://github.com/pgjdbc/pgjdbc/blob/master/ubenchmark/src/main/java/org/postgresql/benchmark/statement/InsertBatch.java#L93-L114

`java -Ddatabase=.. -Duser=.. -jar benchmarks.jar insertBatch -p p1nrows=1024 -prof gc`

## Before
The response times match expectations. When grouping by 2, 4, etc rows, insert gets faster.
I wonder what if use COPY instead.

```
Benchmark                                    (nrows)  (multi)  Mode  Cnt       Score        Error   Units
InsertBatch.insertBatch                         1024        1  avgt   10       8,474 ±      0,240   ms/op
InsertBatch.insertBatch:·gc.alloc.rate.norm     1024        1  avgt   10  569816,413 ±   1987,684    B/op
InsertBatch.insertBatch                         1024        2  avgt   10       5,698 ±      0,726   ms/op
InsertBatch.insertBatch:·gc.alloc.rate.norm     1024        2  avgt   10  403507,417 ±     38,013    B/op
InsertBatch.insertBatch                         1024        4  avgt   10       4,091 ±      0,842   ms/op
InsertBatch.insertBatch:·gc.alloc.rate.norm     1024        4  avgt   10  328752,101 ±     22,011    B/op
InsertBatch.insertBatch                         1024        8  avgt   10       2,979 ±      0,727   ms/op
InsertBatch.insertBatch:·gc.alloc.rate.norm     1024        8  avgt   10  290327,100 ±     25,513    B/op
InsertBatch.insertBatch                         1024       16  avgt   10       2,633 ±      0,560   ms/op
InsertBatch.insertBatch:·gc.alloc.rate.norm     1024       16  avgt   10  271636,664 ±     14,572    B/op
```

## After
The test is performed with `reWriteBatchedInserts=true` property.
It is clear that `reWriteBatchedInserts` is a regression here: number of allocated bytes is *72 times worse*. Response time for "insert 1024 rows with autorewrites" is a bit worse than the one of non-patched version.
```
Benchmark                                    (nrows)  (multi)  Mode  Cnt         Score        Error   Units
InsertBatch.insertBatch                         1024        1  avgt   10         8,942 ±      0,050   ms/op
InsertBatch.insertBatch:·gc.alloc.rate.norm     1024        1  avgt   10  41431294,171 ±   3678,955    B/op
InsertBatch.insertBatch                         1024        2  avgt   10         5,819 ±      1,051   ms/op
InsertBatch.insertBatch:·gc.alloc.rate.norm     1024        2  avgt   10    403507,426 ±     37,764    B/op
InsertBatch.insertBatch                         1024        4  avgt   10         3,949 ±      0,777   ms/op
InsertBatch.insertBatch:·gc.alloc.rate.norm     1024        4  avgt   10    328751,289 ±     22,404    B/op
InsertBatch.insertBatch                         1024        8  avgt   10         2,978 ±      0,499   ms/op
InsertBatch.insertBatch:·gc.alloc.rate.norm     1024        8  avgt   10    290325,533 ±     18,036    B/op
InsertBatch.insertBatch                         1024       16  avgt   10         2,662 ±      0,731   ms/op
InsertBatch.insertBatch:·gc.alloc.rate.norm     1024       16  avgt   10    271636,626 ±     14,678    B/op
```