@sehrope 2016-01-14
-1 to having something like this in the driver itself. There's plenty of better ways to speed up inserts (ex: use COPY) and something like this that rewrites the insert SQL is going to have a lot of issues/corner cases.

One obvious issue is that outside of a transaction, rewriting N sequential INSERT statements as a single INSERT (with multiple values) changes the error handling. Now any error (ex: duplicate key) would prevent any rows from being inserted vs stopping at the first error.