@whitingjr 2016-01-15
@vlsi
> Can you please look into it?
 Yes I will.

> I wonder what if use COPY instead.

 COPY isn't an option for users that are using an ORM. Typically in an application server like Wildfly users are creating JEE applications with Hibernate to handle persistence mapping. It isn't bulk uploading of data.