@vlsi 2016-01-15
>COPY isn't an option for users that are using an ORM.

I mean: "rewrite insert(...) into COPY"