@whitingjr 2016-01-15
@sehrope 
> One obvious issue is that outside of a transaction, rewriting N sequential INSERT statements as a single INSERT (with multiple values) changes the error handling. Now any error (ex: duplicate key) would prevent any rows from being inserted vs stopping at the first error.

 Which is why this feature has a configuration property reWriteBatchedInserts. Defaulted to false.
 If a user is not confident they can handle errors outside a transaction the feature probably should not be used.