@vlsi 2016-01-15
>One obvious issue is that outside of a transaction, rewriting N sequential INSERT statements as a single INSERT (with multiple values) changes the error handling

Technically speaking, some savepoint dance can be implemented to "transparently rollback multi-insert" and "repeat with using regular batching". Thus it might result in the same semantics (except accurate row numbers for each "batched statement") yet the common case is optimized.