@whitingjr 2016-03-15
## prof gc
### Before using 8bee06be4b9679d880f97dbe68cf9a301dd6c420
```
InsertBatch.insertBatch:·gc.alloc.rate                         1024          1  avgt   60      12.163 ±      0.433  MB/sec
InsertBatch.insertBatch:·gc.alloc.rate.norm                    1024          1  avgt   60  545318.463 ±   1477.804    B/op
InsertBatch.insertBatch:·gc.alloc.rate                         1024          4  avgt   60      17.136 ±      0.700  MB/sec
InsertBatch.insertBatch:·gc.alloc.rate.norm                    1024          4  avgt   60  322977.290 ±    786.034    B/op
InsertBatch.insertBatch:·gc.alloc.rate                         1024          8  avgt   60      24.379 ±      1.205  MB/sec
InsertBatch.insertBatch:·gc.alloc.rate.norm                    1024          8  avgt   60  287367.439 ±    295.564    B/op
InsertBatch.insertBatch:·gc.alloc.rate                         1024         16  avgt   60      23.467 ±      2.344  MB/sec
InsertBatch.insertBatch:·gc.alloc.rate.norm                    1024         16  avgt   60  270131.301 ±      8.185    B/op
InsertBatch.insertBatch:·gc.alloc.rate                         1024        128  avgt   60      25.469 ±      2.235  MB/sec
InsertBatch.insertBatch:·gc.alloc.rate.norm                    1024        128  avgt   60  255119.874 ±      6.165    B/op
```
### After using 9198513d16e56b6381a8ecf771fdd89c76542832
```
InsertBatch.insertBatch:·gc.alloc.rate                         1024          1  avgt   60      43.340 ±      2.295  MB/sec
InsertBatch.insertBatch:·gc.alloc.rate.norm                    1024          1  avgt   60  434217.677 ±    709.732    B/op
InsertBatch.insertBatch:·gc.alloc.rate                         1024          4  avgt   60      17.276 ±      0.925  MB/sec
InsertBatch.insertBatch:·gc.alloc.rate.norm                    1024          4  avgt   60  330837.709 ±     12.200    B/op
InsertBatch.insertBatch:·gc.alloc.rate                         1024          8  avgt   60      24.334 ±      1.425  MB/sec
InsertBatch.insertBatch:·gc.alloc.rate.norm                    1024          8  avgt   60  288306.039 ±      7.424    B/op
InsertBatch.insertBatch:·gc.alloc.rate                         1024         16  avgt   60      24.836 ±      1.732  MB/sec
InsertBatch.insertBatch:·gc.alloc.rate.norm                    1024         16  avgt   60  270640.640 ±      6.786    B/op
InsertBatch.insertBatch:·gc.alloc.rate                         1024        128  avgt   60      25.078 ±      2.241  MB/sec
InsertBatch.insertBatch:·gc.alloc.rate.norm                    1024        128  avgt   60  255184.099 ±      6.210    B/op
```
## Response time
### Before
```
Benchmark                (p1nrows)  (p2multi)  Mode  Cnt   Score   Error  Units
InsertBatch.insertBatch         16        128  avgt   60   5.325 ± 0.915  ms/op
InsertBatch.insertBatch        128        128  avgt   60   5.740 ± 0.934  ms/op
InsertBatch.insertBatch       1024          1  avgt   60  45.225 ± 3.958  ms/op
InsertBatch.insertBatch       1024          4  avgt   60  19.473 ± 3.382  ms/op
InsertBatch.insertBatch       1024          8  avgt   60  11.675 ± 1.444  ms/op
InsertBatch.insertBatch       1024         16  avgt   60  14.135 ± 9.557  ms/op
InsertBatch.insertBatch       1024        128  avgt   60  11.010 ± 2.494  ms/op

```
### After
```
Benchmark                (p1nrows)  (p2multi)  Mode  Cnt   Score   Error  Units
InsertBatch.insertBatch         16        128  avgt   60   5.479 ± 1.057  ms/op
InsertBatch.insertBatch        128        128  avgt   60   5.882 ± 0.959  ms/op
InsertBatch.insertBatch       1024          1  avgt   60  11.070 ± 2.211  ms/op
InsertBatch.insertBatch       1024          4  avgt   60  19.550 ± 2.341  ms/op
InsertBatch.insertBatch       1024          8  avgt   60  12.555 ± 2.131  ms/op
InsertBatch.insertBatch       1024         16  avgt   60  11.169 ± 1.220  ms/op
InsertBatch.insertBatch       1024        128  avgt   60  11.071 ± 3.091  ms/op
```